Avoid abbreviations, move MutexTwoPriorities to rclcpp::detail::MutexTwoPriorities
