+++ rclcpp/src/rclcpp/executors/multi_threaded_executor.cpp
@@ -22,6 +22,7 @@
 #include "rclcpp/utilities.hpp"
 #include "rclcpp/scope_exit.hpp"
 
+using rclcpp::detail::MutexTwoPriorities;
 using rclcpp::executors::MultiThreadedExecutor;
 
 MultiThreadedExecutor::MultiThreadedExecutor(
@@ -44,7 +45,6 @@ MultiThreadedExecutor::~MultiThreadedExecutor() {}
 void
 MultiThreadedExecutor::spin()
 {
-  using MutexTwoPriorities = rclcpp::executors::MultiThreadedExecutor::MutexTwoPriorities;
   if (spinning.exchange(true)) {
     throw std::runtime_error("spin() called while already spinning");
   }
@@ -52,8 +52,8 @@ MultiThreadedExecutor::spin()
   std::vector<std::thread> threads;
   size_t thread_id = 0;
   {
-    auto lp_wait_mutex = wait_mutex_.lp();
-    std::lock_guard<MutexTwoPriorities::LpMutex> wait_lock(lp_wait_mutex);
+    auto low_priority_wait_mutex = wait_mutex_.get_low_priority_mutex();
+    std::lock_guard<MutexTwoPriorities::LowPriorityMutex> wait_lock(low_priority_wait_mutex);
     for (; thread_id < number_of_threads_ - 1; ++thread_id) {
       auto func = std::bind(&MultiThreadedExecutor::run, this, thread_id);
       threads.emplace_back(func);
@@ -75,12 +75,11 @@ MultiThreadedExecutor::get_number_of_threads()
 void
 MultiThreadedExecutor::run(size_t)
 {
-  using MutexTwoPriorities = rclcpp::executors::MultiThreadedExecutor::MutexTwoPriorities;
   while (rclcpp::ok(this->context_) && spinning.load()) {
     rclcpp::AnyExecutable any_exec;
     {
-      auto lp_wait_mutex = wait_mutex_.lp();
-      std::lock_guard<MutexTwoPriorities::LpMutex> wait_lock(lp_wait_mutex);
+      auto low_priority_wait_mutex = wait_mutex_.get_low_priority_mutex();
+      std::lock_guard<MutexTwoPriorities::LowPriorityMutex> wait_lock(low_priority_wait_mutex);
       if (!rclcpp::ok(this->context_) || !spinning.load()) {
         return;
       }
@@ -107,8 +106,8 @@ MultiThreadedExecutor::run(size_t)
     execute_any_executable(any_exec);
 
     if (any_exec.timer) {
-      auto hp_wait_mutex = wait_mutex_.hp();
-      std::lock_guard<MutexTwoPriorities::HpMutex> wait_lock(hp_wait_mutex);
+      auto high_priority_wait_mutex = wait_mutex_.get_high_priority_mutex();
+      std::lock_guard<MutexTwoPriorities::HighPriorityMutex> wait_lock(high_priority_wait_mutex);
       auto it = scheduled_timers_.find(any_exec.timer);
       if (it != scheduled_timers_.end()) {
         scheduled_timers_.erase(it);
